+++ lib/dear-inventory-ruby/models/address.rb
@@ -41,28 +41,6 @@ module DearInventoryRuby
     # Points that Address is used as default for chosen Type. `false` as default.
     attr_accessor :default_for_type
 
-    class EnumAttributeValidator
-      attr_reader :datatype
-      attr_reader :allowable_values
-
-      def initialize(datatype, allowable_values)
-        @allowable_values = allowable_values.map do |value|
-          case datatype.to_s
-          when /Integer/i
-            value.to_i
-          when /Float/i
-            value.to_f
-          else
-            value
-          end
-        end
-      end
-
-      def valid?(value)
-        !value || allowable_values.include?(value)
-      end
-    end
-
     # Attribute mapping from ruby-style variable name to JSON key.
     def self.attribute_map
       {
@@ -198,8 +176,6 @@ module DearInventoryRuby
       return false if !@post_code.nil? && @post_code.to_s.length > 20
       return false if @country.nil?
       return false if @type.nil?
-      type_validator = EnumAttributeValidator.new('String', ["Billing", "Business", "Shipping"])
-      return false unless type_validator.valid?(@type)
       true
     end
 
@@ -251,16 +227,6 @@ module DearInventoryRuby
       end
 
       @post_code = post_code
-    end
-
-    # Custom attribute writer method checking allowed values (enum).
-    # @param [Object] type Object to be assigned
-    def type=(type)
-      validator = EnumAttributeValidator.new('String', ["Billing", "Business", "Shipping"])
-      unless validator.valid?(type)
-        fail ArgumentError, "invalid value for \"type\", must be one of #{validator.allowable_values}."
-      end
-      @type = type
     end
 
     # Checks equality by comparing each attribute.